% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{birdweight}
\alias{birdweight}
\title{Weights of birds}
\format{
A matrix with 3 rows and 10 columns.
\itemize{
\item Each row records estimates for a type of bird
(chicken, duck goose)
\item Each column represents one random draw
from the joint distribution of weights.
}
}
\usage{
birdweight
}
\description{
A simulated posterior sample from a model
of bird weights (in kilograms)
}
\keyword{datasets}
