% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-functions.R
\name{check_na_rm}
\alias{check_na_rm}
\title{Check that 'na_rm' is a logial flag}
\usage{
check_na_rm(na_rm)
}
\arguments{
\item{na_rm}{TRUE or FALSE}
}
\value{
TRUE, invisibly
}
\description{
Check that 'na_rm' is a logial flag
}
