% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{dicerolls}
\alias{dicerolls}
\title{Dice rolls}
\format{
A matrix with 6 rows and 10 columns.
\itemize{
\item Each row records outcomes for a particular value.
\item Each column records outcomes for a particular draw.
}
}
\usage{
dicerolls
}
\description{
Data on outcomes from 10 rolls of a dice
}
\keyword{datasets}
