% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{divorce}
\alias{divorce}
\title{Divorce rates in New Zealand}
\format{
A tibble with the following variables:
\itemize{
\item \code{age}: Age, in 5-year age groups, 15-19 to 65+.
\item \code{sex}: \code{"Female"} or \code{"Male"}.
\item \code{time}: Calendar year.
\item \code{sim}: Index for draw.
\item \code{rate}: Divorce rate.
}
}
\source{
Derived from data in tables "Age at divorces by
sex (marriages and civil unions) (Annual-Dec)" and
"Estimated Resident Population by Age and Sex (1991+)
(Annual-Dec)" in the online
database Infoshare
on the Statistics New Zealand website.
Divorce data downloaded on 22 March 2023, and
population data downloaded on 26 March 2023.
}
\usage{
divorce
}
\description{
Posterior sample from a model of divorce rates
in New Zealand.
}
\keyword{datasets}
