% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{n_draw_df}
\alias{n_draw_df}
\title{Given a data frame containing rvecs,
calculate 'n_draw'}
\usage{
n_draw_df(df)
}
\arguments{
\item{df}{A data
the}
}
\description{
If 'n_draw' is the same across all rvecs,
return it. Otherwise raise an error.
}
