% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_rvec}
\alias{new_rvec}
\title{Create a Blank Rvec}
\usage{
new_rvec(x = double(), length = 0, n_draw = 1000)
}
\arguments{
\item{x}{Object with the intended type.
Default is \code{double()}.}

\item{length}{Desired length of rvec.
Default is \code{0}.}

\item{n_draw}{Number of draws of rvec.
Default is \code{1000}.}
}
\value{
An rvec.
}
\description{
Create an rvec, consisting entirely of \code{NAs},
with a given length and number of draws.
}
\details{
The type of the object is taken from \code{x}.
If \code{typeof(x)} is \code{"integer"}, for instance,
then \code{new_rvec()} returns an object
of class \code{"rvec_int"}.
}
\examples{
new_rvec()
new_rvec(TRUE, length = 3, n_draw = 100)

x <- new_rvec(length = 2)
x[1] <- rnorm_rvec(n = 1, n_draw = 1000)
x[2] <- runif_rvec(n = 1, n_draw = 1000)
}
\seealso{
\itemize{
\item \code{\link[=rvec]{rvec()}} \code{\link[=rvec_chr]{rvec_chr()}}, \code{\link[=rvec_dbl]{rvec_dbl()}}, \code{\link[=rvec_int]{rvec_int()}}, \code{\link[=rvec_lgl]{rvec_lgl()}}
Create an rvec from data.
\item \code{\link[=n_draw]{n_draw()}} Query number of draws.
}
}
