% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_rvec.R
\name{reduce_rvec}
\alias{reduce_rvec}
\alias{reduce_rvec.rvec}
\alias{reduce_rvec.rvec_chr}
\alias{reduce_rvec.rvec_lgl}
\title{Scalar summaries of draws}
\usage{
reduce_rvec(x, method = c("median", "mean", "mode"), na_rm = FALSE)

\method{reduce_rvec}{rvec}(x, method = c("median", "mean", "mode"), na_rm = FALSE)

\method{reduce_rvec}{rvec_chr}(x, method = c("median", "mean", "mode"), na_rm = FALSE)

\method{reduce_rvec}{rvec_lgl}(x, method = c("median", "mean", "mode"), na_rm = FALSE)
}
\arguments{
\item{x}{An object of class \link[=rvec]{rvec}.}

\item{method}{Summary measure: a choice of
\code{"median"} (the default), \code{"mean"}, and \code{"mode"}.}

\item{na_rm}{Whether to remove NAs before
calculating summaries. Default is \code{FALSE}.}
}
\value{
A vector.
}
\description{
Reduce an \link[=rvec]{rvec} object to a vector
of scalar summaries.
}
\details{
When \code{method} is \code{"mode"}, \code{reduce_rvec()}
returns the most common value for each
observation. When there is a tie, it returns
\code{NA}.
}
\examples{
m <- cbind(c(1, 1, 1, 2, 3),
           c(2, 4, 0, 2, 3),
           c(0, 0, 1, 0, 100))
x <- rvecs(m)
x
reduce_rvec(x) # defaults to median
reduce_rvec(x, method = "mean")
reduce_rvec(x, method = "mode")
}
\seealso{
Multi-number summaries:
\itemize{
\item \code{\link[=reduce_rvec_ci]{reduce_rvec_ci()}}
\item \code{\link[=reduce_rvec]{reduce_rvec()}}
}

Means and medians are calculated using functions
from package \code{matrixStats}.
}
