% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{regr_coef}
\alias{regr_coef}
\title{Regression coefficients}
\format{
A tibble with 200 rows and
the following variables:
\itemize{
\item \code{draw}: Index identifying random draw.
\item \code{alpha}: Intercept parameter
\item \code{beta}: Slope parameter
\item \code{sigma}: Standard deviation of error term
}
}
\source{
\code{regr_coef} contains values from the second
half of the \code{line} dataset
in package \href{https://CRAN.R-project.org/package=coda}{coda}.
The line dataset draws on the BUGS manual:
Spiegelhalter, D.J., Thomas, A., Best, N.G. and
Gilks, W.R. (1995) BUGS: Bayesian inference using
Gibbs Sampling, Version 0.5, MRC Biostatistics Unit,
Cambridge.
}
\usage{
regr_coef
}
\description{
Posterior sample for coefficients from a linear
regression model.
}
\keyword{datasets}
