% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dexp_rvec}
\alias{dexp_rvec}
\title{The Exponential Distribution, Using Multiple Draws}
\usage{
dexp_rvec(x, rate = 1, log = FALSE)
}
\arguments{
\item{x}{Quantiles. Can be rvec.}

\item{rate}{Vector of rates.
See \code{\link[stats:Exponential]{stats::dexp()}}.
Can be an rvec.}
}
\value{
\itemize{
\item If any of the arguments are rvecs,
or if a value for \code{n_draw} is supplied,
then an \link[=rvec]{rvec}
\item Otherwise an ordinary R vector.
}
}
\description{
Density, distribution function,
quantile function and random generation for the
exponential distribution, modified to work with
rvecs.
}
\details{
Functions \code{dexp_rvec()}, \code{pexp_rvec()},
\code{pexp_rvec()} and \code{rexp_rvec()} work like
base R functions \code{\link[=dexp]{dexp()}}, \code{\link[=pexp]{pexp()}},
\code{\link[=qexp]{qexp()}}, and \code{\link[=rexp]{rexp()}}, except that
they accept rvecs as inputs. If any
input is an rvec, then the output will be too.
Function \code{rexp_rvec()} also returns an
rvec if a value for \code{n_draw} is supplied.

\code{dexp_rvec()}, \code{pexp_rvec()},
\code{pexp_rvec()} and \code{rexp_rvec()}
use \link[vctrs:vector_recycling_rules]{tidyverse}
vector recycling rules:
\itemize{
\item Vectors of length 1 are recycled
\item All other vectors must have the same size
}
}
\examples{
x <- rvec(list(c(3, 5.1),
               c(0.1, 2.3)))
dexp_rvec(x, rate = 1.5)
pexp_rvec(x, rate = 1.5)

rexp_rvec(n = 2,
          rate = c(1.5, 4),
          n_draw = 100)
}
\seealso{
\itemize{
\item \code{\link[=dexp]{dexp()}}
\item \code{\link[=pexp]{pexp()}}
\item \code{\link[=qexp]{qexp()}}
\item \code{\link[=rexp]{rexp()}}
\item \link[stats:Distributions]{stats::distributions}.
}
}
