% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{draws_median}
\alias{draws_median}
\alias{draws_median.rvec_chr}
\alias{draws_median.rvec}
\alias{draws_mean}
\alias{draws_mean.rvec}
\alias{draws_mean.rvec_chr}
\alias{draws_mode}
\alias{draws_mode.rvec}
\title{Medians, means, and modes for random draws}
\usage{
draws_median(x, na_rm = FALSE)

\method{draws_median}{rvec_chr}(x, na_rm = FALSE)

\method{draws_median}{rvec}(x, na_rm = FALSE)

draws_mean(x, na_rm = FALSE)

\method{draws_mean}{rvec}(x, na_rm = FALSE)

\method{draws_mean}{rvec_chr}(x, na_rm = FALSE)

draws_mode(x, na_rm = FALSE)

\method{draws_mode}{rvec}(x, na_rm = FALSE)
}
\arguments{
\item{x}{An object of class \link[=rvec]{rvec}.}

\item{na_rm}{Whether to remove NAs before
calculating summaries. Default is \code{FALSE}.}
}
\value{
A vector.
}
\description{
Summarise the distribution of random draws
in an \code{rvec}, using means, medians, or modes.
}
\details{
An \code{rvec}\tabular{lcccc}{
    \tab Draw 1 \tab Draw 2 \tab \eqn{\dots} \tab Draw \eqn{n} \cr
   Obs 1 \tab \eqn{x_{1,1}} \tab \eqn{x_{1,2}} \tab \eqn{\dots} \tab \eqn{x_{1,n}} \cr
   Obs 2 \tab \eqn{x_{2,1}} \tab \eqn{x_{2,2}} \tab \eqn{\dots} \tab \eqn{x_{2,n}} \cr
   \eqn{\vdots} \tab \eqn{\vdots} \tab \eqn{\vdots} \tab \eqn{\ddots} \tab \eqn{\vdots} \cr
   Obs \eqn{m} \tab \eqn{x_{m,1}} \tab \eqn{x_{m,2}} \tab \eqn{\dots} \tab \eqn{x_{m,n}} \cr
}


When \code{method} is \code{"mode"}, \code{reduce_rvec()}
returns the most common value for each
observation. When there is a tie, it returns
\code{NA}.
}
\examples{
m <- rbind(a = c(1, 1, 1, 2, 3),
           b = c(2, 4, 0, 2, 3),
           c = c(0, 0, 1, 0, 100))
x <- rvec(m)
x
draws_median(x)
draws_mean(x)
draws_mode(x)
}
\seealso{
\itemize{
\item \code{\link[=draws_quantile]{draws_quantile()}} to calculate quantiles
\item \code{\link[=draws_fun]{draws_fun()}} to apply abritrary functions
}

For a wide range of tools for summarising random draws, see
\href{https://CRAN.R-project.org/package=tidybayes}{tidybayes}
and \href{https://CRAN.R-project.org/package=ggdist}{ggdist}.
Function \code{\link[=as_list_col]{as_list_col()}} converts \code{rvec}s into a
format that \code{tidybayes} and \code{ggdist} can work with.
}
