% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{draws_quantile}
\alias{draws_quantile}
\alias{draws_quantile.rvec}
\alias{draws_quantile.rvec_chr}
\title{Quantiles for random draws}
\usage{
draws_quantile(x, probs = c(0.025, 0.5, 0.975), na_rm = FALSE)

\method{draws_quantile}{rvec}(x, probs = c(0.025, 0.5, 0.975), na_rm = FALSE)

\method{draws_quantile}{rvec_chr}(x, probs = c(0.025, 0.5, 0.975), na_rm = FALSE)
}
\arguments{
\item{x}{An object of class \link[=rvec]{rvec}.}

\item{probs}{Vector of probabilities.}

\item{na_rm}{Whether to remove NAs before
calculating summaries. Default is \code{FALSE}.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Summarise the distribution of random draws
in an \code{rvec}, using quantiles.
}
\details{
The \code{probs} argument defaults to
\code{c(0.025, 0.5, 0.975)}, the values needed
for a median and 95\% credible interval.

Regular \link[base:data.frame]{data frames}
and \link[tibble:tibble]{tibbles} both allow
data frame columns, so the results from
\code{draws_quantile()} can be inserted into a
data frame. See below for an example.

To expand a data frame column, use
\href{https://tidyr.tidyverse.org/reference/hoist.html}{tidyr::unnest_wider()}
}
\examples{
set.seed(0)
m <- rbind(a = rnorm(100, mean = 5, sd = 2),
           b = rnorm(100, mean = -3, sd = 3),
           c = rnorm(100, mean = 0, sd = 20))
x <- rvec(m)
x
draws_quantile(x)

## results from 'draw_quantile'
## assigned to a data frame
library(dplyr)
df <- tibble(x)
df \%>\%
  mutate(draws_quantile(x))
}
\seealso{
\itemize{
\item \code{\link[=draws_median]{draws_median()}}, \link{draws_mean}, \code{\link[=draws_mode]{draws_mode()}}
describe the location of a distribution.
\item \code{\link[=draws_fun]{draws_fun()}} applies a function to draws.
}
}
