% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_rvec}
\alias{is_rvec}
\title{Is an Object an Rvec}
\usage{
is_rvec(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Test whether \code{x} inherits from
class \code{"rvec"}.
}
\examples{
x <- rvec_dbl()
is_rvec(x)
}
\seealso{
\itemize{
\item \code{\link[=rvec]{rvec()}} to create an rvec
\item \code{\link[=as.matrix]{as.matrix()}}, \code{\link[=as_list_col]{as_list_col()}},
to convert an rvec into other formats
}
}
