% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{missing}
\alias{missing}
\alias{anyNA.rvec}
\alias{is.na.rvec}
\alias{na.exclude.rvec}
\alias{na.omit.rvec}
\title{Missing, Finite, and Infinite Values
in Rvecs}
\usage{
\method{anyNA}{rvec}(x, recursive = FALSE)

\method{is.na}{rvec}(x)

\method{na.exclude}{rvec}(object, ...)

\method{na.omit}{rvec}(object, ...)
}
\arguments{
\item{x, object}{An \link[=rvec]{rvec}.}

\item{recursive}{Whether \code{anyNA()} should be
applied recursively to lists. Ignored when
\code{x} is an rvec.}

\item{...}{Currently ignored.}
}
\description{
rvec methods for \code{is.na()}, \code{is.nan()},
\code{is.finite()}, and \code{is.infinite()} all
return rvecs, with separate values for
every draw.

The rvec method for \code{anyNA()} returns
\code{TRUE} if any value for any draw is
\code{NA} or \code{NaN}, and returns \code{FALSE} otherwise.

Functions \code{na.omit()} and \code{na.exclude()}
return versions of \code{x} in which observations
with \code{NA} in any draws are omitted.
}
\details{
The behavior of the rvec methods
for \code{is.na()}, \code{is.nan()},
\code{is.finite()}, and \code{is.infinite()}
differs from the standard
\href{https://vctrs.r-lib.org}{vctrs}
behavior, which is to return a logical
vector with length equal to \code{length(x)}.
With rvecs, the standard \strong{vctrs} behavior
would entail summarising across draws,
which is the job of the \link[=draws_all]{draws_*}
functions.
}
\examples{
x <- rvec(list(c(1.2, NA),
               c(Inf, 3),
               c(-1, NaN)))

## return a logical rvec
is.na(x)
is.nan(x)
is.finite(x)
is.infinite(x)

## return a logical scalar
anyNA(x)

## return an NA-free version of 'x'
na.omit(x)
na.exclude(x)

## use 'if_else_rvec' to modify values
## within rvec
if_else_rvec(is.na(x), 999, x)

## vctrs functions
library(vctrs, warn.conflicts = FALSE)
## all draws missing
vec_detect_missing(x)
## any draws missing
vec_detect_complete(x)
}
\seealso{
\itemize{
\item \code{\link[=if_else_rvec]{if_else_rvec()}} for modifying individual
values within draws.
\item Base R functions \code{\link[=is.na]{is.na()}}, \code{\link[=is.nan]{is.nan()}},
\code{\link[=is.finite]{is.finite()}}, \code{\link[=is.infinite]{is.infinite()}},
\code{\link[=anyNA]{anyNA()}}, \code{\link[=na.omit]{na.omit()}}, \code{\link[=na.exclude]{na.exclude()}}
\item \code{\link[vctrs:missing]{vctrs::vec_detect_missing()}} to test whether
all draws for an observation are missing.
\item \code{\link[vctrs:vec_detect_complete]{vctrs::vec_detect_complete()}} to test whether
any draws for an observation are missing.
}
}
