% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-functions.R
\name{missing}
\alias{missing}
\title{Missing and infinite values}
\description{
\link[vctrs:missing]{is.na} in vctrs: a row of a data frame
is NA only if all the elements are missing.
}
\details{
\code{is.finite()}, \code{is.infinite()}, \code{is.nan()} interpreted the same way
}
\examples{
## 'is.na' operates on entire sets of draws,
## and is only TRUE of all draws are NA
m1 <- cbind(c(1, 2, NA),
            c(4, 5, 6),
            c(NA, NA, NA))
x1 <- rvec(m1)
is.na(x1)

## 'is.nan' works the same way
m2 <- cbind(c(1, 2, NaN),
            c(4, 5, 6),
            c(NaN, NaN, NaN))
x2 <- rvec(m2)
is.nan(x2)

## and so do 'is.finite' and 'is.infinite'
m3 <- cbind(c(1, 2, Inf),
            c(4, 5, 6),
            c(Inf, Inf, Inf))
x3 <- rvec(m2)
is.finite(x3)
is.infinite(x3)
}
\seealso{
\itemize{
\item \code{\link[vctrs:missing]{vctrs::vec_detect_missing()}}
\item \code{\link[vctrs:vec_detect_complete]{vctrs::vec_detect_complete()}}
\item \code{\link[vctrs:missing]{vctrs::vec_any_missing()}}
}
}
