% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rvec-distributions}
\alias{rvec-distributions}
\alias{dnorm_rvec}
\alias{pnorm_rvec}
\alias{qnorm_rvec}
\alias{rnorm_rvec}
\alias{dpois_rvec}
\alias{ppois_rvec}
\alias{qpois_rvec}
\alias{rpois_rvec}
\title{Probability distributions}
\usage{
dnorm_rvec(x, mean = 0, sd = 1, log = FALSE)

pnorm_rvec(q, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

qnorm_rvec(p, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

rnorm_rvec(n, mean = 0, sd = 1)

dpois_rvec(x, lambda, log = FALSE)

ppois_rvec(q, lambda, lower.tail = TRUE, log.p = FALSE)

qpois_rvec(p, lambda, lower.tail = TRUE, log.p = FALSE)

rpois_rvec(n, lambda)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{mean, lambda}{Vector of means.
A standard R vector, or an \link[=rvec]{rvec}.}

\item{sd}{Vector of standard deviations.
A standard R vector, or an \link[=rvec]{rvec}.}

\item{log, log.p}{Whether to return
\code{log(p)} rather than \code{p}. Default is
\code{FALSE}.}

\item{lower.tail}{Whether to return
\eqn{P[X \le x]}, as opposed to
\eqn{P[X > x]}. Default is \code{TRUE}.}

\item{p}{Vector of probabilities.}

\item{n}{Number of draws.}
}
\value{
If any of the inputs is an
\link[=rvec]{rvec}, then the return value is
an \link[=rvec]{rvec}. Otherwise it is a
standard R vector.
}
\description{
Modified versions of standard probability
functions that take one or more rvecs
as arguments, and return rvecs.
}
\details{
Standard base R rules
about recycling of arguments apply.
For instance,

\if{html}{\out{<div class="sourceCode">}}\preformatted{rnorm_rvec(n = 4,
           mean = rvec(rbind(c(0.3, 0.7),
                             c(0.2, 0.5))))
}\if{html}{\out{</div>}}

is equivalent to

\if{html}{\out{<div class="sourceCode">}}\preformatted{rnorm_rvec(n = 4,
           mean = rvec(rbind(c(0.3, 0.7),
                             c(0.2, 0.5),
                             c(0.3, 0.7),
                             c(0.2, 0.5))))
}\if{html}{\out{</div>}}

with the \code{mean} argument being repeated so
that it has length 4.
}
\examples{
x_rv <- rvec(rbind(c(-0.8, 1.3),
                   c(-9.1, 8.7)))
mean_rv <- rvec(rbind(c(-1, 1),
                      c(-10, 10)))
sd_rv <- rvec(rbind(c(0.2, 20)))
x_rv
mean_rv
sd_rv

## densities: all arguments rvecs
dnorm_rvec(x = x_rv, mean = mean_rv, sd = sd_rv)

## densities: 'x' is ordinary vector
dnorm_rvec(x = c(0, 2), mean = mean_rv, sd = sd_rv)

## ...which is equivalent to
c(dnorm_rvec(0, mean = mean_rv[1], sd = sd_rv),
  dnorm_rvec(2, mean = mean_rv[2], sd = sd_rv))
      
## random variates: mean is rvec, sd is rvec 
rnorm_rvec(n = 2, mean = mean_rv, sd = sd_rv)

## random variates: mean is rvec, sd is scalar
rnorm_rvec(n = 6, mean = mean_rv, sd = 0.5)
}
\seealso{
Base R functions
\itemize{
\item \code{\link[stats:Normal]{stats::dnorm()}} Normal distribution.
\item \code{\link[stats:Poisson]{stats::dpois()}} Poisson distribution.
}
}
