% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rvec-distributions}
\alias{rvec-distributions}
\alias{dbeta_rvec}
\alias{pbeta_rvec}
\alias{qbeta_rvec}
\alias{rbeta_rvec}
\alias{dbinom_rvec}
\alias{pbinom_rvec}
\alias{qbinom_rvec}
\alias{rbinom_rvec}
\alias{dcauchy_rvec}
\alias{pcauchy_rvec}
\alias{qcauchy_rvec}
\alias{rcauchy_rvec}
\alias{dnorm_rvec}
\alias{pnorm_rvec}
\alias{qnorm_rvec}
\alias{rnorm_rvec}
\alias{dpois_rvec}
\alias{ppois_rvec}
\alias{qpois_rvec}
\alias{rpois_rvec}
\title{Probability distributions}
\usage{
dbeta_rvec(x, shape1, shape2, ncp = 0, log = FALSE)

pbeta_rvec(q, shape1, shape2, ncp = 0, lower.tail = TRUE, log.p = FALSE)

qbeta_rvec(p, shape1, shape2, ncp = 0, lower.tail = TRUE, log.p = FALSE)

rbeta_rvec(n, shape1, shape2, ncp = 0)

dbinom_rvec(x, size, prob, log = FALSE)

pbinom_rvec(q, size, prob, lower.tail = TRUE, log.p = FALSE)

qbinom_rvec(p, size, prob, lower.tail = TRUE, log.p = FALSE)

rbinom_rvec(n, size, prob)

dcauchy_rvec(x, location = 0, scale = 1, log = FALSE)

pcauchy_rvec(q, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

qcauchy_rvec(p, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

rcauchy_rvec(n, location = 0, scale = 1)

dnorm_rvec(x, mean = 0, sd = 1, log = FALSE)

pnorm_rvec(q, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

qnorm_rvec(p, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

rnorm_rvec(n, mean = 0, sd = 1)

dpois_rvec(x, lambda, log = FALSE)

ppois_rvec(q, lambda, lower.tail = TRUE, log.p = FALSE)

qpois_rvec(p, lambda, lower.tail = TRUE, log.p = FALSE)

rpois_rvec(n, lambda)
}
\arguments{
\item{x, q}{Vector of quantiles. Can be rvec.}

\item{ncp}{Non-centrality parameter. Cannot be rvec.
Default is 0.}

\item{log, log.p}{Whether to return
\code{log(p)} rather than \code{p}. Cannot be rvec. Default is
\code{FALSE}.}

\item{lower.tail}{Whether to return
\eqn{P[X \le x]}, as opposed to
\eqn{P[X > x]}. Cannot be rvec. Default is \code{TRUE}.}

\item{p}{Vector of probabilities. Can be rvec.}

\item{n}{Number of draws. Cannot be rvec.}

\item{mean}{Vector of means. Can be rvec.
Default is 0.}

\item{sd}{Vector of standard deviations. Can be rvec.
Default is 1.}

\item{lambda}{Vector of means. Can be rvec.}
}
\value{
If any of the inputs is an
\link[=rvec]{rvec}, then an
an \link[=rvec]{rvec}; otherwise an
ordinary R vector.
}
\description{
Modified versions of standard probability
functions that take one or more rvecs
as arguments, and return rvecs.
}
\details{
The \verb{*_rvec} probability distribution functions
use \link[vctrs:vector_recycling_rules]{tidyverse}
vector recycling rules:
\itemize{
\item Vectors of length 1 are recycled
\item All other vectors must have the same size
}

These rules are more restrictive than base R rules,
but are also more predictable. Base R style
recycling can be done through explicit
calls to \code{\link[base:rep]{base::rep()}},
\code{\link[base:rep]{base::rep_len()}}, and \code{\link[base:rep]{base::rep.int()}},
all of which have methods for \link[=rvec]{rvecs}.
}
\examples{
x_rv <- rvec(list(c(-0.8, 1.3),
                  c(-9.1, 8.7)))
mean_rv <- rvec(list(c(-1, 1),
                     c(-10, 10)))
sd_rv <- rvec(list(c(0.2, 20)))
x_rv
mean_rv
sd_rv

## densities: all arguments rvecs
dnorm_rvec(x = x_rv, mean = mean_rv, sd = sd_rv)

## densities: 'x' is ordinary vector
dnorm_rvec(x = c(0, 2), mean = mean_rv, sd = sd_rv)

## ...which is equivalent to
c(dnorm_rvec(0, mean = mean_rv[1], sd = sd_rv),
  dnorm_rvec(2, mean = mean_rv[2], sd = sd_rv))

## random variates: mean is rvec, sd is rvec
rnorm_rvec(n = 2, mean = mean_rv, sd = sd_rv)

## random variates: mean is rvec, sd is scalar
rnorm_rvec(n = 2, mean = mean_rv, sd = 0.5)
}
\seealso{
Base R functions
\itemize{
\item \code{\link[stats:Beta]{stats::dbeta()}} Beta distribution.
\item \code{\link[stats:Binomial]{stats::dbinom()}} Binomial distribution.
\item \code{\link[stats:Cauchy]{stats::dcauchy()}} Cauchy distribution.
\item \code{\link[stats:Chisquare]{stats::dchisq()}} Chi-squared distribution.
\item \code{\link[stats:Normal]{stats::dnorm()}} Normal distribution.
\item \code{\link[stats:Poisson]{stats::dpois()}} Poisson distribution.
}
}
