% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOps.R
\docType{methods}
\name{rvec-matrix-mult}
\alias{rvec-matrix-mult}
\alias{\%*\%,Matrix,rvec-method}
\alias{\%*\%,rvec,Matrix-method}
\title{Internal Methods for Matrix Multiplication with rvec}
\usage{
\S4method{\%*\%}{Matrix,rvec}(x, y)

\S4method{\%*\%}{rvec,Matrix}(x, y)
}
\description{
These methods support \code{\%*\%} between Matrix and rvec objects.
}
\keyword{internal}
