% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvec-package.R
\docType{package}
\name{rvec-package}
\alias{rvec-package}
\title{Package 'rvec'}
\description{
Many modern statistical methods, including Bayesian methods,
yield draws from a distribution for the quantities of interest.
An enormously useful feature of draws from a distribution is
that they can used to generate distributions for new
quantities. Posterior distributions for death rates, for instance,
can be turned into posterior distributions for life expectancies.
}
\details{
Starting with a sample

\deqn{x_1, x_2, \dots, x_n}

we apply one or more functions to each draw,
to obtain a sample

\deqn{f(g(x_1)), f(g(x_2)), \dots, f(g(x_n))},

from the distribution of the  quantity \eqn{f(g(x))}.
This derived distribution can, like any distribution,
be summarised through indicators such as means and
quantiles.

Package \code{rvec} makes  recipe easy to follow.
}
\section{Standard workflow}{


\strong{Step 1: Create an \code{rvec}}

Create an \code{rvec} object holding draws from a posterior sample.
\itemize{
\item \code{\link[=rvec]{rvec()}} - class depends on input
\item \code{\link[=rvec_dbl]{rvec_dbl()}} - doubles
\item \code{\link[=rvec_int]{rvec_int()}} - integers
\item \code{\link[=rvec_lgl]{rvec_lgl()}} - logical
\item \code{\link[=rvec_chr]{rvec_chr()}} - character
}

\strong{Step 2: Manipulate the \code{rvec}}

Perform calculations on the \code{rvec}, treating it like an ordinary vector.
Internally, the calculations the manipulations are applied independenly
to each draw.
\itemize{
\item Standard mathematical and statistical operations
\item Special rvec operations
}

\strong{Step 3: Summarise the \code{rvec}}
\itemize{
\item \code{\link[=draw_median]{draw_median()}}
\item \code{\link[=draw_mean]{draw_mean()}}
\item \code{\link[=draw_mode]{draw_mode()}}
\item \code{\link[=draw_quantile]{draw_quantile()}}
}
}

\section{Other functions}{


\link{collapse_to_rvec}
}

\section{Design}{


\code{rvec} are built from \link[vctrs:vctrs-package]{vctrs} package, which
look after many of the tricky details, and give it a sound
conceptual basis. Internally, many of the calculations are
done using the \link{matrixStats} package, is fast.
}

\section{Other packages}{


posterior: different type of data structure,
but also lots of diagnostics

ggdist

distributional
}

\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

}
