% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvec-package.R
\docType{package}
\name{rvec-package}
\alias{rvec-package}
\title{Package 'rvec'}
\description{
\itemize{
\item logic from \code{vctrs}
\item uses \code{matrixStats} for speed
}
}
\details{
\itemize{
\item motivation
\itemize{
\item reduce size of data frames
\item make
}
}
}
\section{Functions}{
\subsection{Creating and converting}{

Create individual \code{rvec}s:
\itemize{
\item \code{\link[=rvec]{rvec()}}
\item \code{\link[=rvec_dbl]{rvec_dbl()}}
\item \code{\link[=rvec_int]{rvec_int()}}
\item \code{\link[=rvec_lgl]{rvec_lgl()}}
\item \code{\link[=rvec_chr]{rvec_chr()}}
}

Reformat an entire data frame:
\itemize{
\item \code{\link[=collapse_to_rvec]{collapse_to_rvec()}}
\item \code{\link[=expand_from_rvec]{expand_from_rvec()}}
}

Casting to other formats:
\itemize{
\item \code{\link[=as_list_col]{as_list_col()}}
\item \code{\link[=as_rvec]{as_rvec()}}
\item \code{\link[=as.matrix]{as.matrix()}}
\item \code{\link[=as_rvar]{as_rvar()}}
}
}

\subsection{Map operations}{
\itemize{
\item vctrs vs base vectors
\item \code{\link[=if_else_rvec]{if_else_rvec()}}
}
}

\subsection{Reduce operations}{

\code{\link[=reduce_rvec]{reduce_rvec()}}
\code{\link[=reduce_rvec_cols]{reduce_rvec_cols()}}
}
}

\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

}
