% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvec-constructors.R
\name{rvec}
\alias{rvec}
\alias{rvec_chr}
\alias{rvec_dbl}
\alias{rvec_int}
\alias{rvec_lgl}
\title{Create a random vector}
\usage{
rvec(x)

rvec_chr(x = NULL)

rvec_dbl(x = NULL)

rvec_int(x = NULL)

rvec_lgl(x = NULL)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A random vector with the following class:
\itemize{
\item \code{rvec_dbl()}: \code{"rvec_dbl"}
\item \code{rvec_int()}: \code{"rvec_int"}
\item \code{rvec_lgl()}: \code{"rvec_lgl"}
\item \code{rvec_chr()}: \code{"rvec_chr"}
\item \code{rvec()}: \code{"rvec_chr"}, \code{"rvec_dbl"}
\code{"rvec_int"}, or \code{"rvec_lgl"}, depending on
\code{typeof(x)}.
}
}
\description{
Create a vector of class \code{"rvec"}. Class
\code{"rvec"} has four subclasses, each dealing with
a diffent type:
\itemize{
\item \code{"rvec_dbl"} doubles
\item \code{"rvec_int"} integers
\item \code{"rvec_lgl"} logical
\item \code{"rvec_chr"} character
}

Argument \code{x} is a matrix in which each
row represents a set of draws for a particular
quantity.
}
\examples{
birdweight
rvec_dbl(birdweight)
rvec(birdweight)

dicerolls
rvec(dicerolls)
}
