% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvec-constructors.R
\name{rvec}
\alias{rvec}
\alias{rvec_chr}
\alias{rvec_dbl}
\alias{rvec_int}
\alias{rvec_lgl}
\title{Create an rvec}
\usage{
rvec(x)

rvec_chr(x = NULL)

rvec_dbl(x = NULL)

rvec_int(x = NULL)

rvec_lgl(x = NULL)
}
\arguments{
\item{x}{A matrix or vector.}
}
\value{
An rvec with the following class:
\itemize{
\item \code{rvec_dbl()}: \code{"rvec_dbl"}
\item \code{rvec_int()}: \code{"rvec_int"}
\item \code{rvec_lgl()}: \code{"rvec_lgl"}
\item \code{rvec_chr()}: \code{"rvec_chr"}
\item \code{rvec()}: \code{"rvec_chr"}, \code{"rvec_dbl"}
\code{"rvec_int"}, or \code{"rvec_lgl"}, depending on
\code{typeof(x)}.
}
}
\description{
Create an object of class \code{"rvec"}.
}
\details{
Class \code{"rvec"} has four subclasses, each dealing with
a diffent type:
\itemize{
\item \code{"rvec_dbl"} doubles
\item \code{"rvec_int"} integers
\item \code{"rvec_lgl"} logical
\item \code{"rvec_chr"} character
}

These subclasses are analogous to \code{\link[=double]{double()}},
\code{\link[=integer]{integer()}}, \code{\link[=logical]{logical()}}, and \code{\link[=character]{character()}}
vectors.

Function \code{rvec()} chooses the subclass, based on
\code{x}. Functions \code{rvec_dbl()}, \code{rvec_int()},
\code{rvec_lgl()}, and \code{rvec_chr()} each create
objects of a particular subclass.

\code{x} is typically a matrix, where each row is
a set of draws for a particular unknown quantity.
However, \code{x} can also be a vector, which is
treated as a single-row matrix. \code{x} can be
doubles, integers, logical, or character.
}
\examples{
birdweight
rvec_dbl(birdweight)
rvec(birdweight)

dicerolls
rvec(dicerolls)
}
